+++ up42/__init__.py
@@ -28,7 +28,7 @@ from up42.order_template import BatchOrderTemplate
 from up42.processing import Job, JobSorting, JobStatus
 from up42.stac import BulkDeletion
 from up42.stac import extend as stac_extend
-from up42.tasking import FeasibilityStudy, FeasibilityStudySorting, Quotation, QuotationSorting
+from up42.tasking import FeasibilityStudy, FeasibilityStudySorting, OrderCoverage, Quotation, QuotationSorting
 from up42.utils import get_up42_py_version
 from up42.version import version_control
 
@@ -60,5 +60,6 @@ __all__ = [
         FeasibilityStudy,
         FeasibilityStudySorting,
         BulkDeletion,
+        OrderCoverage,
     ]
 ]