+++ CHANGELOG.md
@@ -29,6 +29,10 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+### 3.3.0a1
+**February 10, 2026**
+### Added
+- Added functionality to `tasking.py` get order's coverage
 
 ### 3.2.0
 **February 2, 2026**